add: clients for each element
