+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cotalker-sdk",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Cotalker's partners lib",
   "main": "lib/index.js",
   "author": "cotalker",
