+++ src/customTypes/COTTypes/APIGenerics.ts
@@ -1,0 +1,6 @@
+export type isActiveOptions = 'true'|'false'|'all'  //archiuvo apigenerics.ts
+
+export type JSONPatchBody = {
+  op: 'add'|'remove'|'replace'|'move'|'copy'|'test'
+  path: `/${string}`, value?: unknown
+}[]  //apigenerics
