+++ src/utils/HttpClient.ts
@@ -46,26 +46,25 @@ export default abstract class HttpClient {
 
   protected _handleError = async (axiosError: AxiosError): Promise<any> => {
     const { config, response } = axiosError
-    if (response.status === 429) {
-      console.info(`429 TOO MANY REQUESTS: Retrying ${config.method} ${config.url}`) 
-      const waitTime = Number(response.headers['retry-after'] ?? response.headers['Retry-After'])*1000
-      return new Promise(resolve=>setTimeout(resolve, waitTime))
-      .then(function () {
+    switch (response.status) {
+      case 429:
+        console.error(`Retrying ${config.method} ${config.url}`)  
+        const waitTime = Number(response.headers['retry-after'] ?? response.headers['Retry-After'])*1000
+        await new Promise(resolve=>setTimeout(resolve, waitTime))
         return (axios(config))
-      })
-    } else
-    if (response.status === 401) {
-      console.info(`401 UNAUTHORIZED: unvalid token`)
-    } else
-    if (response.status === 403) {
-      console.info(`403 FORBIDDEN: User has no permission to take this action`)
-    } else
-    if (response.status === 404) {
-      console.info(`404 NOT FOUND£`)
-    } else
-    if (response.status === 500) {
-      console.info(`500 INTERNAL SERVER ERROR: `)
-    }
+      case 401:
+        throw new Error("Bad authorization token. Your token has expired.")
+      case 403:
+        throw new Error("You do not have permissions to do this action! ")
+      case 404:
+        console.error("ID not found", config.url)
+        return Promise.resolve({data: null})
+      case 500:
+        console.error(" Cotalker API replied with internal error", response.data)
+        return Promise.resolve({data: null})
+      default:
+        break;
+    } 
     return Promise.reject(axiosError);
   }
 }
